package name changed to com.mdg.androble
