+++ app/build.gradle
@@ -2,10 +2,10 @@ apply plugin: 'com.android.application'
 
 android {
     compileSdkVersion 24
-    buildToolsVersion "24.0.0"
+    buildToolsVersion "24.0.1"
 
     defaultConfig {
-        applicationId "com.example.thispc.testlibrary"
+        applicationId "com.mdg.androble"
         minSdkVersion 15
         targetSdkVersion 24
         versionCode 2
@@ -23,5 +23,5 @@ dependencies {
     compile fileTree(dir: 'libs', include: ['*.jar'])
     testCompile 'junit:junit:4.12'
     compile 'com.android.support:appcompat-v7:24.1.1'
-    compile project(":bluetooth_library2")
+    compile project(':library')
 }
