+++ app/src/main/res/layout/activity_main.xml
@@ -8,7 +8,7 @@
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
     android:background="#ffffff"
-    tools:context="com.example.thispc.testlibrary.MainActivity">
+    tools:context="com.mdg.androble.testlibrary.com.mdg.androble.MainActivity">
 
     <Button
         android:layout_width="wrap_content"
