--- bluetooth_library2/build.gradle
@@ -1,110 +1,0 @@
-apply plugin: 'com.android.library'
-
-android {
-    compileSdkVersion 24
-    buildToolsVersion "24.0.0"
-
-    defaultConfig {
-        minSdkVersion 15
-        targetSdkVersion 24
-        versionCode 3
-        versionName "0.1.1"
-    }
-    buildTypes {
-        release {
-            minifyEnabled false
-            proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
-        }
-    }
-}
-
-dependencies {
-    compile fileTree(dir: 'libs', include: ['*.jar'])
-    testCompile 'junit:junit:4.12'
-    compile 'com.android.support:appcompat-v7:24.1.1'
-}
-
-
-apply plugin: 'com.github.dcendents.android-maven'
-apply plugin: 'com.jfrog.bintray'
-
-version = "0.1.0"
-def siteUrl = 'https://github.com/pkarira/Androble'    // project homepage
-def gitUrl = 'https://github.com/pkarira/Androble.git' // project git
-group = "com.sdsmdg.androble"
-
-install {
-    repositories.mavenInstaller {
-        // This generates POM.xml with proper parameters
-        pom {
-            project {
-                packaging 'aar'
-                name 'Library For Android'
-                url siteUrl
-                licenses {
-                    license {
-                        name 'The Apache software License, Version 2.0'
-                        url 'http://www.apache.org/licenses/LICENSE-2.0.txt'
-                    }
-                }
-                developers {
-                    developer {
-                        id ''
-                        name 'Rahul Yadav'
-                        email 'rahuly247@gmail.com'
-                    }
-                }
-                scm {
-                    connection gitUrl
-                    developerConnection gitUrl
-                    url siteUrl
-                }
-            }
-        }
-    }
-}
-task sourcesJar(type: Jar) {
-    from android.sourceSets.main.java.srcDirs
-    classifier = 'sources'
-}
-
-task javadoc(type: Javadoc) {
-    source = android.sourceSets.main.java.srcDirs
-    classpath += project.files(android.getBootClasspath().join(File.pathSeparator))
-}
-
-task javadocJar(type: Jar, dependsOn: javadoc) {
-    classifier = 'javadoc'
-    from javadoc.destinationDir
-}
-
-javadoc {
-    options{
-        encoding 'UTF-8'
-        charSet 'UTF-8'
-        author true
-    }
-}
-
-artifacts {
-    archives javadocJar
-    archives sourcesJar
-}
-
-
-Properties properties = new Properties()
-properties.load(project.rootProject.file('local.properties').newDataInputStream())
-bintray {
-    user = properties.getProperty("bintray.user")
-    key = properties.getProperty("bintray.apikey")
-    configurations = ['archives']
-    pkg {
-        repo = "maven"
-        name = "androble"                // project name in maven
-        websiteUrl = siteUrl
-        vcsUrl = gitUrl
-        licenses = ["Apache-2.0"]
-        publish = true
-    }
-}
-
